Begin to fix unit testing: Vue Menu.
