+++ tests/unit/components/menu/ColumnMenuItem-test.js
@@ -1,16 +1,18 @@
 "use strict"
 
 import { mount } from '@test'
-import ArtefactMenuItem from '~/components/ArtefactMenuItem.vue'
+import ColumnMenuItem from '~/components/menu/ColumnMenuItem.vue'
 
-describe("ArtefactMenuItem", function() {
+describe("ColumnMenuItem", function() {
     let wrapper, vm
     const fakeID = 12345
+    const name = 'test col'
 
     beforeEach(() => {
-        wrapper = mount(ArtefactMenuItem, {
+        wrapper = mount(ColumnMenuItem, {
             propsData: {
                 dataId: fakeID,
+                name: name,
             }
         })
         vm = wrapper.vm
@@ -18,18 +20,19 @@ describe("ArtefactMenuItem", function() {
     
     it('pushes data to the router', () => {
         const push = sinon.spy()
-        const artID = 12345
+        const colID = 12345
 
         // create wrapper with mocked route and routers
-        let wrapper = mount(ArtefactMenuItem, {
+        let wrapper = mount(ColumnMenuItem, {
             propsData: {
-                artefact: artID,
+                dataId: colID,
+                name: name,
             },
             mocks: { 
                 $router: { push },
                 $route: {
                     params: { 
-                        artID,
+                        colID,
                         scrollID: 808,
                         scrollVersionID: 1,
                     }
@@ -40,7 +43,7 @@ describe("ArtefactMenuItem", function() {
 
         // assertions
         expect(push.firstCall.args[0].name).to.equal("workbenchAddress")
-        expect(push.firstCall.args[0].params).to.include({ artID })
+        expect(push.firstCall.args[0].params).to.include({ colID })
     })
 
     it('has a span', () => {
